refactor(core): cleanup code and add documentation
