+++ src/core/religions/Faith.cpp
@@ -2,9 +2,12 @@
 
 Faith::Faith() {}
 
-Faith::Faith(const std::string& name, const sf::Color& color)
-: m_Name(name), m_Color(color)
+Faith::Faith(const std::string& name, const sf::Color& color) :
+    m_Name(name),
+    m_Color(color)
 {}
+
+//////////////////////////////////////////////////////
 
 std::string Faith::GetName() const {
     return m_Name;
@@ -14,6 +17,8 @@ sf::Color Faith::GetColor() const {
     return m_Color;
 }
 
+//////////////////////////////////////////////////////
+
 void Faith::SetName(const std::string& name) {
     m_Name = name;
 }
@@ -21,3 +26,5 @@ void Faith::SetName(const std::string& name) {
 void Faith::SetColor(const sf::Color& color) {
     m_Color = color;
 }
+
+//////////////////////////////////////////////////////
