+++ src/core/titles/Title.cpp
@@ -1,8 +1,9 @@
 #include "Title.hpp"
 #include "mod/Mod.hpp"
-#include "provinces/ProvinceManager.hpp"
 
-Title::Title() : Title("", sf::Color(0, 0, 0)) {}
+Title::Title() :
+    Title("", sf::Color(0, 0, 0))
+{}
 
 Title::Title(std::string name, sf::Color color, bool landless) :
     m_Name(name),
@@ -13,7 +14,46 @@ Title::Title(std::string name, sf::Color color, bool landless) :
     m_SelectionFocus(true)
 {}
 
-// Title::Title(const Title& title) : Title(title.GetName(), title.GetColor()) {}
+//////////////////////////////////////////////////////
+
+bool Title::Is(TitleType type) const {
+    return this->GetType() == type;
+}
+
+bool Title::IsVassal(HighTitle* title) const {
+    if (title == nullptr)
+        return false;
+    HighTitle* liege = m_LiegeTitle;
+    while(liege != nullptr) {
+        if(liege == title)
+            return true;
+        liege = liege->GetLiegeTitle();
+    }
+    return false;
+}
+
+bool Title::HasLocName(const std::string& lang) const {
+    auto it = m_LocNames.find(lang);
+    if (it == m_LocNames.end())
+        return false;
+    return !it->second.empty();
+}
+
+bool Title::HasLocAdjective(const std::string& lang) const {
+    auto it = m_LocAdjectives.find(lang);
+    if (it == m_LocAdjectives.end())
+        return false;
+    return !it->second.empty();
+}
+
+bool Title::HasLocArticle(const std::string& lang) const {
+    auto it = m_LocArticles.find(lang);
+    if (it == m_LocArticles.end())
+        return false;
+    return !it->second.empty();
+}
+
+//////////////////////////////////////////////////////
 
 std::string Title::GetName() const {
     return m_Name;
@@ -39,38 +79,6 @@ bool Title::IsLandless() const {
     return m_Landless;
 }
 
-bool Title::Is(TitleType type) const {
-    return this->GetType() == type;
-}
-
-bool Title::IsVassal(HighTitle* title) const {
-    if (title == nullptr)
-        return false;
-    HighTitle* liege = m_LiegeTitle;
-    while(liege != nullptr) {
-        if(liege == title)
-            return true;
-        liege = liege->GetLiegeTitle();
-    }
-    return false;
-}
-
-void Title::SetName(std::string name) {
-    m_Name = name;
-}
-
-void Title::SetColor(sf::Color color) {
-    m_Color = color;
-}
-
-void Title::SetLiegeTitle(HighTitle* title) {
-    m_LiegeTitle = title;
-}
-
-void Title::SetLandless(bool landless) {
-    m_Landless = landless;
-}
-
 std::string Title::GetOriginalFileName() const {
     return m_OriginalFileName;
 }
@@ -79,44 +87,16 @@ SharedPtr<Jomini::Object> Title::GetOriginalData() const {
     return m_OriginalData;
 }
 
-void Title::SetOriginalFileName(const std::string& fileName) {
-    m_OriginalFileName = fileName;
-}
-
-void Title::SetOriginalData(SharedPtr<Jomini::Object> data) {
-    m_OriginalData = data;
-}
-
 std::string Title::GetOriginalHistoryFileName() const {
     return m_OriginalHistoryFileName;
-}
-
-void Title::SetOriginalHistoryFileName(const std::string& fileName) {
-    m_OriginalHistoryFileName = fileName;
 }
 
 std::map<Jomini::Date, SharedPtr<Jomini::Object>>& Title::GetHistory() {
     return m_History;
 }
 
-void Title::AddHistory(Jomini::Date date, SharedPtr<Jomini::Object> data) {
-    m_History[date] = data;
-}
-
-void Title::RemoveHistory(Jomini::Date date) {
-    m_History.erase(date);
-}
-
 std::map<std::string, std::string>& Title::GetCulturalNames() {
     return m_CulturalNames;
-}
-
-void Title::AddCulturalName(const std::string& culture, std::string name) {
-    m_CulturalNames[culture] = name;
-}
-
-void Title::RemoveCulturalName(const std::string& culture) {
-    m_CulturalNames.erase(culture);
 }
 
 std::map<std::string, std::string>& Title::GetLocNames() {
@@ -134,17 +114,6 @@ std::string Title::GetLocName(const std::string& lang) const {
     return it->second;
 }
 
-bool Title::HasLocName(const std::string& lang) const {
-    auto it = m_LocNames.find(lang);
-    if (it == m_LocNames.end())
-        return false;
-    return !it->second.empty();
-}
-
-void Title::SetLocName(const std::string& lang, std::string name) {
-    m_LocNames[lang] = name;
-}
-
 std::map<std::string, std::string>& Title::GetLocAdjectives() {
     return m_LocAdjectives;
 }
@@ -157,17 +126,6 @@ std::string Title::GetLocAdjective(const std::string& lang) const {
     if(m_LocAdjectives.count(lang) == 0)
         return "";
     return m_LocAdjectives.at(lang);
-}
-
-bool Title::HasLocAdjective(const std::string& lang) const {
-    auto it = m_LocAdjectives.find(lang);
-    if (it == m_LocAdjectives.end())
-        return false;
-    return !it->second.empty();
-}
-
-void Title::SetLocAdjective(const std::string& lang, std::string adjective) {
-    m_LocAdjectives[lang] = adjective;
 }
 
 std::map<std::string, std::string>& Title::GetLocArticles() {
@@ -184,146 +142,72 @@ std::string Title::GetLocArticle(const std::string& lang) const {
     return m_LocArticles.at(lang);
 }
 
-bool Title::HasLocArticle(const std::string& lang) const {
-    auto it = m_LocArticles.find(lang);
-    if (it == m_LocArticles.end())
-        return false;
-    return !it->second.empty();
+bool Title::HasSelectionFocus() const {
+    return m_SelectionFocus;
+}
+
+//////////////////////////////////////////////////////
+
+void Title::SetName(std::string name) {
+    m_Name = name;
+}
+
+void Title::SetColor(sf::Color color) {
+    m_Color = color;
+}
+
+void Title::SetLiegeTitle(HighTitle* title) {
+    m_LiegeTitle = title;
+}
+
+void Title::SetLandless(bool landless) {
+    m_Landless = landless;
+}
+
+void Title::SetOriginalFileName(const std::string& fileName) {
+    m_OriginalFileName = fileName;
+}
+
+void Title::SetOriginalData(SharedPtr<Jomini::Object> data) {
+    m_OriginalData = data;
+}
+
+void Title::SetOriginalHistoryFileName(const std::string& fileName) {
+    m_OriginalHistoryFileName = fileName;
+}
+
+void Title::SetLocName(const std::string& lang, std::string name) {
+    m_LocNames[lang] = name;
+}
+
+void Title::SetLocAdjective(const std::string& lang, std::string adjective) {
+    m_LocAdjectives[lang] = adjective;
 }
 
 void Title::SetLocArticle(const std::string& lang, std::string article) {
     m_LocArticles[lang] = article;
 }
 
-bool Title::HasSelectionFocus() const {
-    return m_SelectionFocus;
-}
-
 void Title::SetSelectionFocus(bool focus) {
     m_SelectionFocus = focus;
 }
 
-HighTitle::HighTitle() : Title("", sf::Color(0, 0, 0)), m_CapitalTitle(nullptr) {}
+//////////////////////////////////////////////////////
 
-HighTitle::HighTitle(std::string name, sf::Color color, bool landless) : Title(name, color, landless), m_CapitalTitle(nullptr) {}
-
-std::vector<Title*>& HighTitle::GetDejureTitles() {
-    return m_DejureTitles;
+void Title::AddHistory(Jomini::Date date, SharedPtr<Jomini::Object> data) {
+    m_History[date] = data;
 }
 
-const std::vector<Title*>& HighTitle::GetDejureTitles() const {
-    return m_DejureTitles;
+void Title::RemoveHistory(Jomini::Date date) {
+    m_History.erase(date);
 }
 
-CountyTitle* HighTitle::GetCapitalTitle() {
-    return m_CapitalTitle;
+void Title::AddCulturalName(const std::string& culture, std::string name) {
+    m_CulturalNames[culture] = name;
 }
 
-bool HighTitle::HasDejureTitle(const Title* title) const {
-    return std::find(m_DejureTitles.begin(), m_DejureTitles.end(), title) != m_DejureTitles.end();
+void Title::RemoveCulturalName(const std::string& culture) {
+    m_CulturalNames.erase(culture);
 }
 
-void HighTitle::AddDejureTitle(Title* title) {
-    if (!this->HasDejureTitle(title)) {
-        m_DejureTitles.push_back(title);
-
-        HighTitle* previousLiege = title->GetLiegeTitle();
-        if (previousLiege != nullptr) {
-            previousLiege->RemoveDejureTitle(title);
-        }
-    }
-    title->SetLiegeTitle(this);
-}
-
-void HighTitle::RemoveDejureTitle(Title* title) {
-    m_DejureTitles.erase(
-        std::remove(m_DejureTitles.begin(), m_DejureTitles.end(), title),
-        m_DejureTitles.end()
-    );
-    title->SetLiegeTitle(nullptr);
-}
-
-void HighTitle::SetCapitalTitle(CountyTitle* title) {
-    m_CapitalTitle = title;
-}
-
-void HighTitle::ClearDejureTitles() {
-    for (Title* dejure : m_DejureTitles)
-        dejure->SetLiegeTitle(nullptr);
-    m_DejureTitles.clear();
-}
-
-void HighTitle::SetSelectionFocus(bool focus) {
-    m_SelectionFocus = focus;
-    if(focus) {
-        for(Title* dejureTitle : m_DejureTitles)
-            dejureTitle->SetSelectionFocus(true);
-    }
-}
-
-sf::Vector2i HighTitle::GetImagePosition(const ProvinceManager& provinceManager) const {
-    if(m_DejureTitles.empty())
-        return sf::Vector2i(0, 0);
-    return m_DejureTitles.front()->GetImagePosition(provinceManager);
-}
-
-BaronyTitle::BaronyTitle() : Title(), m_ProvinceId(0) {}
-BaronyTitle::BaronyTitle(std::string name, sf::Color color, bool landless) : Title(name, color, landless), m_ProvinceId(0) {}
-BaronyTitle::BaronyTitle(std::string name, sf::Color color, bool landless, int provinceId) : Title(name, color, landless), m_ProvinceId(provinceId) {}
-
-TitleType BaronyTitle::GetType() const {
-    return TitleType::BARONY;
-}
-
-int BaronyTitle::GetProvinceId() const {
-    return m_ProvinceId;
-}
-
-void BaronyTitle::SetProvinceId(int id) {
-    m_ProvinceId = id;
-}
-
-bool BaronyTitle::HasSelectionFocus() const {
-    return true;
-}
-
-sf::Vector2i BaronyTitle::GetImagePosition(const ProvinceManager& provinceManager) const {
-    if (const Province* province = provinceManager.GetProvinceById(m_ProvinceId))
-        return province->GetImagePosition();
-    return sf::Vector2i(0, 0);
-}
-
-CountyTitle::CountyTitle() : HighTitle() {}
-CountyTitle::CountyTitle(std::string name, sf::Color color, bool landless) : HighTitle(name, color, landless) {}
-
-TitleType CountyTitle::GetType() const {
-    return TitleType::COUNTY;
-}
-
-DuchyTitle::DuchyTitle() : HighTitle() {}
-DuchyTitle::DuchyTitle(std::string name, sf::Color color, bool landless) : HighTitle(name, color, landless) {}
-
-TitleType DuchyTitle::GetType() const {
-    return TitleType::DUCHY;
-}
-
-KingdomTitle::KingdomTitle() : HighTitle() {}
-KingdomTitle::KingdomTitle(std::string name, sf::Color color, bool landless) : HighTitle(name, color, landless) {}
-
-TitleType KingdomTitle::GetType() const {
-    return TitleType::KINGDOM;
-}
-
-EmpireTitle::EmpireTitle() : HighTitle() {}
-EmpireTitle::EmpireTitle(std::string name, sf::Color color, bool landless) : HighTitle(name, color, landless) {}
-
-TitleType EmpireTitle::GetType() const {
-    return TitleType::EMPIRE;
-}
-
-HegemonyTitle::HegemonyTitle() : HighTitle() {}
-HegemonyTitle::HegemonyTitle(std::string name, sf::Color color, bool landless) : HighTitle(name, color, landless) {}
-
-TitleType HegemonyTitle::GetType() const {
-    return TitleType::HEGEMONY;
-}
+//////////////////////////////////////////////////////
