+++ src/core/titles/TitleType.hpp
@@ -13,6 +13,11 @@ enum class TitleType {
 const std::vector<const char*> TitleTypeLabels = { "Barony", "County", "Duchy", "Kingdom", "Empire", "Hegemony" };
 const std::vector<const char*> TitleTypePrefixes = { "b", "c", "d", "k", "e", "h" };
 
+/**
+ * @brief Gets the title type by its name.
+ * @param name The name of the title.
+ * @return The title type.
+ */
 inline TitleType GetTitleTypeByName(const std::string& name) {
     if (name.size() > 2 && name[1] == '_' && name[2] != ' ') {
         for(int i = 0; i < static_cast<int>(TitleType::COUNT); i++) {
@@ -23,10 +28,21 @@ inline TitleType GetTitleTypeByName(const std::string& name) {
     throw std::invalid_argument("GetTitleTypeByName: invalid title name");
 }
 
+/**
+ * @brief Gets the prefix for a title type.
+ * @param type The title type.
+ * @return The prefix for the title type.
+ */
 inline std::string GetTitlePrefixByType(TitleType type) {
     return TitleTypePrefixes[static_cast<int>(type)];
 }
 
+/**
+ * @brief Checks if a title name is valid for a specific title type.
+ * @param name The title name to check.
+ * @param type The title type.
+ * @return True if the title name is valid for the specified title type, false otherwise.
+ */
 inline bool IsValidTitleName(const std::string& name, TitleType type) {
     if (name.size() < 3)
         return false;
@@ -39,6 +55,11 @@ inline bool IsValidTitleName(const std::string& name, TitleType type) {
     return TitleTypePrefixes[static_cast<int>(type)][0] == name[0];
 }
 
+/**
+ * @brief Checks if a title name is valid.
+ * @param name The title name to check.
+ * @return True if the title name is valid, false otherwise.
+ */
 inline bool IsValidTitleName(const std::string& name) {
     if (name.size() < 3)
         return false;
